web/backend/requirements.txt
Django==4.0.1
django-grappelli
django-hashedfilenamestorage
django-filter
django-nested-admin
djangorestframework
drf-yasg
django-admin-sortable2